Moved all control variables from int to float

+++ P5/rc.h
@@ -27,7 +27,7 @@ typedef struct
     unsigned char state;
     unsigned int time;
     unsigned short pulse;
-    int setpoint;
+    float setpoint;
 }
 channel;
